MR-48: Added styles
