+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^14.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.9.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
@@ -46,8 +47,8 @@
     "eslint-config-prettier": "^8.8.0",
     "eslint-config-standard": "latest",
     "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-n": "^15.7.0",
+    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "latest",
