+++ src/Components/SuperAdmins/Modal/super-admins.module.css
@@ -1,0 +1,40 @@
+.modalContainer {
+  color:#373867;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modalContent {
+  background-color: #fff;
+  padding: 20px;
+  border-radius: 4px;
+}
+
+.modalButtons {
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 20px;
+}
+
+.modalButton {
+  font-size: 16px;
+  font-weight: 700;
+  border-style: none;
+  height: 40px;
+  width: 60px;
+  background: #AACE9B;
+  border-radius: 5px;
+  margin-left: 20px;
+}
+
+.modalButton:hover {
+  background-color: #2c2b2b;
+  color: white;
+}
