+++ src/Components/SuperAdmins/Table/super-admins.module.css
@@ -6,5 +6,24 @@
 }
 
 .thead {
+  font-size: 2.2vh;
   background-color: #EBECF0;
+}
+
+.tableSuperAdmin {
+  width: 60%;
+  font-size: 2vh;
+}
+
+.tableRow:hover {
+  background-color: #EBECF0;
+}
+
+.deleteButton {
+  color: red;
+}
+
+.actions {
+  text-align: center;
+  font-size: 2vh;
 }
